Fixed warnings and parameter name mismatches, added #include needed in some platforms.

+++ modules/calib3d/src/homography_decomp.cpp
@@ -47,6 +47,7 @@
  //M*/
 
 #include "precomp.hpp"
+#include <memory>
 
 namespace cv
 {
@@ -189,7 +190,7 @@ void HomographyDecompZhang::decompose(std::vector<CameraMotion>& camMotions)
     double lambda1t3 = lambda1*lambda3;
 
     double t1 = 1.0/(2.0*lambda1t3);
-    double t2 = sqrtf(1.0+4.0*lambda1t3/lambda1m3_2);
+    double t2 = sqrt(1.0+4.0*lambda1t3/lambda1m3_2);
     double t12 = t1*t2;
 
     double e1 = -t1 + t12; //t1*(-1.0f + t2 );
@@ -197,8 +198,8 @@ void HomographyDecompZhang::decompose(std::vector<CameraMotion>& camMotions)
     double e1_2 = e1*e1;
     double e3_2 = e3*e3;
 
-    double nv1p = sqrtf(e1_2*lambda1m3_2 + 2*e1*(lambda1t3-1) + 1.0);
-    double nv3p = sqrtf(e3_2*lambda1m3_2 + 2*e3*(lambda1t3-1) + 1.0);
+    double nv1p = sqrt(e1_2*lambda1m3_2 + 2*e1*(lambda1t3-1) + 1.0);
+    double nv3p = sqrt(e3_2*lambda1m3_2 + 2*e3*(lambda1t3-1) + 1.0);
     double v1p[3], v3p[3];
 
     v1p[0]=Vt.at<double>(0)*nv1p, v1p[1]=Vt.at<double>(1)*nv1p, v1p[2]=Vt.at<double>(2)*nv1p;
@@ -378,7 +379,7 @@ void HomographyDecompInria::decompose(std::vector<CameraMotion>& camMotions)
     }
 
     double traceS = S(0, 0) + S(1, 1) + S(2, 2);
-    double v = 2.0 * sqrtf(1 + traceS - M00 - M11 - M22);
+    double v = 2.0 * sqrt(1 + traceS - M00 - M11 - M22);
 
     double ESii = signd(S(indx, indx)) ;
     double r_2 = 2 + traceS + v;
